@@ -1,6 +1,6 @@
 /*
  * *
- *  * Copyright 2021 eBay Inc.
+ *  * Copyright 2024 eBay Inc.
  *  *
  *  * Licensed under the Apache License, Version 2.0 (the "License");
  *  * you may not use this file except in compliance with the License.
@@ -21,29 +21,38 @@
 
 'use strict';
 
-const express = require('express');
+import express, { Request, Response } from 'express';
+import { Config } from '../lib/types/Config';
+import configData from './config.json';
+import * as constants from '../lib/constants';
+import * as EventNotificationSDK from '../lib/index';
+
 const app = express();
+const config: Config = configData as Config;
 
-const config = require('./config.json');
-const constants = require('../lib/constants');
-const EventNotificationSDK = require('../lib/index');
-
-const environment = 'PRODUCTION';
+const environment = constants.ENVIRONMENT.PRODUCTION;
 const PORT = process.env.PORT || 8080;
 
 app.use(express.json());
 
-app.post('/webhook', (req, res) => {
+app.post('/webhook', (req: Request, res: Response) => {
+    const signature = req.headers[constants.X_EBAY_SIGNATURE] as string | undefined;
+
+    if (!signature) {
+        res.status(constants.HTTP_STATUS_CODE.PRECONDITION_FAILED).send('Signature is missing');
+        return;
+    }
+
     EventNotificationSDK.process(
         req.body,
-        req.headers[constants.X_EBAY_SIGNATURE],
+        signature,
         config,
         environment
-    ).then((responseCode) => {
+    ).then((responseCode: number) => {
         if (responseCode === constants.HTTP_STATUS_CODE.NO_CONTENT) {
             console.log(`Message processed successfully for: \n- Topic: ${req.body.metadata.topic} \n- NotificationId: ${req.body.notification.notificationId}\n`);
         } else if (responseCode === constants.HTTP_STATUS_CODE.PRECONDITION_FAILED) {
-            console.error(`Signature mismatch for: \n- Payload: ${JSON.stringify(req.body)} \n- Signature: ${req.headers[constants.X_EBAY_SIGNATURE]}\n`);
+            console.error(`Signature mismatch for: \n- Payload: ${JSON.stringify(req.body)} \n- Signature: ${signature}\n`);
         }
         res.status(responseCode).send();
     }).catch((ex) => {
@@ -52,17 +61,18 @@ app.post('/webhook', (req, res) => {
     });
 });
 
-app.get('/webhook', (req, res) => {
-    if (req.query.challenge_code) {
+app.get('/webhook', (req: Request, res: Response) => {
+    const challengeCode = req.query.challenge_code as string | undefined;
+
+    if (challengeCode) {
         try {
             const challengeResponse = EventNotificationSDK.validateEndpoint(
-                req.query.challenge_code,
+                challengeCode,
                 config);
             res.status(200).send({
                 challengeResponse: challengeResponse
             });
         } catch (e) {
-            // eslint-disable-next-line no-console
             console.error(`Endpoint validation failure: ${e}`);
             res.status(constants.HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR).send();
         }
@@ -72,6 +82,5 @@ app.get('/webhook', (req, res) => {
 });
 
 app.listen(PORT, () => {
-    // eslint-disable-next-line no-console
     console.log(`Listening at http://localhost:${PORT}`);
 });
