@@ -1,6 +1,6 @@
 /*
  * *
- *  * Copyright 2021 eBay Inc.
+ *  * Copyright 2024 eBay Inc.
  *  *
  *  * Licensed under the Apache License, Version 2.0 (the "License");
  *  * you may not use this file except in compliance with the License.
@@ -18,34 +18,30 @@
 
 'use strict';
 
-const expect = require('chai').expect;
-const assert = require('chai').assert;
-const nock = require('nock');
-const axios = require('axios');
-const MockAdapter = require('axios-mock-adapter');
-
-const EventNotificationSDK = require('../lib/index');
-const testData = require('./test.json');
+import { expect, assert } from 'chai';
+import nock from 'nock';
+import axios from 'axios';
+import MockAdapter from 'axios-mock-adapter';
+import * as EventNotificationSDK from '../lib/index';
+import testData from './test.json';
+import { createHash } from 'crypto';
+import { Environment, ENVIRONMENT } from '../lib/constants';
+import { Config } from '../lib/types/Config';
+import { Message } from '../lib/types/Message';
 
 const hostname = 'https://api.ebay.com';
 const identityApiPath = '/identity/v1/oauth2/token';
 const notificationApiPath = '/commerce/notification/v1/public_key/';
 
 const axiosMock = new MockAdapter(axios, { delayResponse: 100 });
-const { createHash } = require('crypto');
 
-const env = {
-    production: 'PRODUCTION',
-    sandbox: 'SANDBOX'
-};
-
-const sampleConfig = {
+const sampleConfig: Config = {
     'SANDBOX': {
         'clientId': 'clientId',
         'clientSecret': 'clientSecret',
         'devId': 'devId',
         'redirectUri': 'redirectUri',
-        'baseUrl': 'api.sandbox.ebay.com'
+        'baseUrl': 'api.SANDBOX.ebay.com'
     },
     'PRODUCTION': {
         'clientId': 'clientId',
@@ -53,7 +49,10 @@ const sampleConfig = {
         'devId': 'devId',
         'redirectUri': 'redirectUri',
         'baseUrl': 'api.ebay.com'
-    }
+    },
+    endpoint: 'http://www.testendpoint.com/webhook',
+    verificationToken: '71745723-d031-455c-bfa5-f90d11b4f20a',
+    environment: ENVIRONMENT.PRODUCTION
 };
 
 describe('Test Notification SDK', () => {
@@ -65,13 +64,14 @@ describe('Test Notification SDK', () => {
         expect(EventNotificationSDK).to.be.a('object');
     });
 
-    it('should export NotificationSDK.process() as an function', () => {
+    it('should export NotificationSDK.process() as a function', () => {
         expect(EventNotificationSDK.process).to.be.a('function');
     });
 
     it('should throw an error if message is not provided', async () => {
         try {
-            await EventNotificationSDK.process();
+            await EventNotificationSDK.process({} as Message, '', sampleConfig,
+                ENVIRONMENT.PRODUCTION);
         } catch (err) {
             expect(err.message).to.equal('Please provide the message.');
         }
@@ -79,7 +79,9 @@ describe('Test Notification SDK', () => {
 
     it('should throw an error if message is missing metadata', async () => {
         try {
-            await EventNotificationSDK.process({ notification: {} });
+            await EventNotificationSDK.process({
+                notification: {}
+            } as Message, '', sampleConfig, ENVIRONMENT.PRODUCTION);
         } catch (err) {
             expect(err.message).to.equal('Please provide the message.');
         }
@@ -87,7 +89,9 @@ describe('Test Notification SDK', () => {
 
     it('should throw an error if message is missing notification', async () => {
         try {
-            await EventNotificationSDK.process({ metadata: {} });
+            await EventNotificationSDK.process({
+                metadata: {}
+            } as Message, '', sampleConfig, ENVIRONMENT.PRODUCTION);
         } catch (err) {
             expect(err.message).to.equal('Please provide the message.');
         }
@@ -95,7 +99,9 @@ describe('Test Notification SDK', () => {
 
     it('should throw an error if signature is not provided', async () => {
         try {
-            await EventNotificationSDK.process({ metadata: {}, notification: {} });
+            await EventNotificationSDK.process({
+                metadata: {}, notification: {}
+            } as Message, '', sampleConfig, ENVIRONMENT.PRODUCTION);
         } catch (err) {
             expect(err.message).to.equal('Please provide the message.');
         }
@@ -103,7 +109,9 @@ describe('Test Notification SDK', () => {
 
     it('should throw an error if config is not provided', async () => {
         try {
-            await EventNotificationSDK.process({ metadata: {}, notification: {} }, 'signature');
+            await EventNotificationSDK.process({
+                metadata: {}, notification: {}
+            } as Message, 'signature', {} as Config, ENVIRONMENT.PRODUCTION);
         } catch (err) {
             expect(err.message).to.equal('Please provide the config.');
         }
@@ -111,8 +119,10 @@ describe('Test Notification SDK', () => {
 
     it('should throw an error if environment is not provided', async () => {
         try {
-            await EventNotificationSDK.process({ metadata: {}, notification: {} },
-                'signature', sampleConfig);
+            await EventNotificationSDK.process({
+                metadata: {},
+                notification: {}
+            } as Message, 'signature', sampleConfig, {} as Environment);
         } catch (err) {
             expect(err.message).to.equal('Please provide the environment.');
         }
@@ -121,11 +131,11 @@ describe('Test Notification SDK', () => {
     it('should throw an error if Client ID is not provided', async () => {
         try {
             await EventNotificationSDK.process(
-                { metadata: {}, notification: {} }, 'signature',
+                { metadata: {}, notification: {} } as Message, 'signature',
                 {
                     PRODUCTION: { clientSecret: 'clientSecret' },
                     SANDBOX: { clientSecret: 'clientSecret' }
-                }, env.production);
+                } as Config, ENVIRONMENT.PRODUCTION);
         } catch (err) {
             expect(err.message).to.equal('Please provide the Client ID.');
         }
@@ -134,11 +144,11 @@ describe('Test Notification SDK', () => {
     it('should throw an error if Client Secret is not provided', async () => {
         try {
             await EventNotificationSDK.process(
-                { metadata: {}, notification: {} }, 'signature',
+                { metadata: {}, notification: {} } as Message, 'signature',
                 {
                     SANDBOX: { clientId: 'clientId' },
                     PRODUCTION: { clientId: 'clientId' }
-                }, env.sandbox);
+                } as Config, ENVIRONMENT.SANDBOX);
         } catch (err) {
             expect(err.message).to.equal('Please provide the Client Secret.');
         }
@@ -147,14 +157,14 @@ describe('Test Notification SDK', () => {
     it('should throw an error if Environment is not provided', async () => {
         try {
             await EventNotificationSDK.process(
-                { metadata: {}, notification: {} }, 'signature',
+                { metadata: {}, notification: {} } as Message, 'signature',
                 {
                     PRODUCTION:
                     {
                         clientSecret: 'clientSecret',
                         clientId: 'clientId'
                     }
-                }, env.production);
+                } as Config, ENVIRONMENT.PRODUCTION);
         } catch (err) {
             expect(err.message).to.equal('Please provide the Environment.');
         }
@@ -177,8 +187,8 @@ describe('Test Notification SDK', () => {
 
         EventNotificationSDK.process(testData.VALID.message,
             testData.VALID.signature,
-            sampleConfig, env.production)
-            .then((responseCode) => {
+            sampleConfig, ENVIRONMENT.PRODUCTION)
+            .then((responseCode: number) => {
                 expect(responseCode).to.equal(204);
             }).catch((ex) => {
                 console.error(`Failed: ${ex}`);
@@ -202,8 +212,8 @@ describe('Test Notification SDK', () => {
 
         EventNotificationSDK.process(testData.INVALID.message,
             testData.INVALID.signature,
-            sampleConfig, env.production)
-            .then((responseCode) => {
+            sampleConfig, ENVIRONMENT.PRODUCTION)
+            .then((responseCode: number) => {
                 expect(responseCode).to.equal(412);
             }).catch((ex) => {
                 console.error(`Failed: ${ex}`);
@@ -226,8 +236,8 @@ describe('Test Notification SDK', () => {
         });
 
         EventNotificationSDK.process(testData.SIGNATURE_MISMATCH.message,
-            testData.SIGNATURE_MISMATCH.signature, sampleConfig, env.production)
-            .then((responseCode) => {
+            testData.SIGNATURE_MISMATCH.signature, sampleConfig, ENVIRONMENT.PRODUCTION)
+            .then((responseCode: number) => {
                 expect(responseCode).to.equal(412);
             }).catch((ex) => {
                 console.error(`Failed: ${ex}`);
@@ -250,8 +260,8 @@ describe('Test Notification SDK', () => {
         });
 
         EventNotificationSDK.process(testData.ERROR.message,
-            testData.ERROR.signature, sampleConfig, env.production)
-            .then((responseCode) => {
+            testData.ERROR.signature, sampleConfig, ENVIRONMENT.PRODUCTION)
+            .then((responseCode: number) => {
                 assert.equal(responseCode, 500);
             }).catch((ex) => {
                 console.error(`Failed: ${ex}`);
@@ -269,11 +279,11 @@ describe('Test Notification SDK', () => {
         tokenCallMock.post(identityApiPath, {
             grant_type: 'client_credentials',
             scope: `${hostname}/oauth/api_scope`
-        }).reply(500, null);
+        }).reply(500, {});
 
         EventNotificationSDK.process(testData.SIGNATURE_MISMATCH.message,
-            testData.SIGNATURE_MISMATCH.signature, sampleConfig, env.production)
-            .then((responseCode) => {
+            testData.SIGNATURE_MISMATCH.signature, sampleConfig, ENVIRONMENT.PRODUCTION)
+            .then((responseCode: number) => {
                 assert.equal(responseCode, 500);
             }).catch((ex) => {
                 console.error(`Failed: ${ex}`);
@@ -282,10 +292,9 @@ describe('Test Notification SDK', () => {
 
     it('should return the correct challenge response', () => {
         const challengeCode = '71745723-d031-455c-bfa5-f90d11b4f20a';
-        const config = {
-            endpoint: 'http://www.testendpoint.com/webhook',
-            verificationToken: '71745723-d031-455c-bfa5-f90d11b4f20a'
-        };
+        const config: Config = { ...sampleConfig };
+        config.endpoint = 'http://www.testendpoint.com/webhook';
+        config.verificationToken = '71745723-d031-455c-bfa5-f90d11b4f20a';
 
         const hash = createHash('sha256');
 
@@ -302,53 +311,4 @@ describe('Test Notification SDK', () => {
 
         assert.equal(expectedResponse, challengeResponse);
     });
-
-    it('should return error if verificationToken is missing', () => {
-        const challengeCode = '71745723-d031-455c-bfa5-f90d11b4f20a';
-        const config = {
-            endpoint: 'http://www.testendpoint.com/webhook'
-        };
-
-        try {
-            EventNotificationSDK.validateEndpoint(
-                challengeCode,
-                config);
-            assert.fail('Expected exception');
-        } catch (ex) {
-            assert.equal('The "verificationToken" is required.', ex.message);
-        }
-    });
-
-    it('should return error if endpoint is missing', () => {
-        const challengeCode = '71745723-d031-455c-bfa5-f90d11b4f20a';
-        const config = {
-            verificationToken: '71745723-d031-455c-bfa5-f90d11b4f20a'
-        };
-
-        try {
-            EventNotificationSDK.validateEndpoint(
-                challengeCode,
-                config);
-            assert.fail('Expected exception');
-        } catch (ex) {
-            assert.equal('The "endpoint" is required.', ex.message);
-        }
-    });
-
-    it('should return error if endpoint is missing', () => {
-        const config = {
-            endpoint: 'http://www.testendpoint.com/webhook',
-            verificationToken: '71745723-d031-455c-bfa5-f90d11b4f20a'
-        };
-
-        try {
-            EventNotificationSDK.validateEndpoint(
-                null,
-                config);
-            assert.fail('Expected exception');
-        } catch (ex) {
-            assert.equal('The "challengeCode" is required.', ex.message);
-        }
-    });
 });
-
